pre-release autofix

+++ src/lib/misc/misc.c
@@ -14,7 +14,7 @@
 /*FIXME rename this file it contains only net_utils function*/
 unsigned char *ipaddr_scan(unsigned char *addr, unsigned char *res) {
 	char symbol_str[4];
-	int i,j, tmp, cur = 0;
+	size_t i, j, tmp, cur = 0;
 	for (i = 0; i < (sizeof(res) - 1); i++) {
 		symbol_str[0]='\0';
 		for (j = 0; j < ARRAY_SIZE(symbol_str); j++) {
@@ -47,7 +47,7 @@ unsigned char *ipaddr_scan(unsigned char *addr, unsigned char *res) {
 
 unsigned char *macaddr_scan(unsigned char *addr, unsigned char *res) {
 	char symbol_str[4];
-	size_t i,j, cur = 0;
+	size_t i, j, cur = 0;
 	unsigned int tmp;
 	for (i = 0; i < 5; i++) {
 		symbol_str[0]='\0';
